+++ app/models/utility/scenario/units/definitions.rb
@@ -352,6 +352,7 @@ module Utility
               "amc_35": { sz: 3, mv: 5, wpn: :"47mm_sa_35", bd: 4, arm: [2, 2, 2], trt: [2, 2, 2] },
               "char_d2": { sz: 4, mv: 4, wpn: :"47mm_sa_35", arm: [3, 3, 3], trt: [3, 3, 3] },
               "somua_s35": { sz: 4, mv: 4, wpn: :"47mm_sa_35", arm: [4, 3, 3], trt: [3, 3, 3] },
+              "renault_ft": { sz: 3, mv: 3, wpn: :"37mm_sa_18", arm: [1, 1, 1], trt: [1, 1, 1] },
               "amr_35_zt3": { sz: 3, mv: 6, wpn: :"25mm_hotchkiss", arm: [1, 1, 1] },
               "amc_schneider_p16": { sz: 3, mv: 6, wpn: :"37mm_sa_18", trn: 1, arm: [1, 1, 1], trt: [1, 1, 1] },
               "white_am_ac": { sz: 3, mv: 4, wpn: :"37mm_sa_18", whl: true, trn: 1, arm: [0, 0, 0], trt: [0, 0, 0] },